proxy: manual buffer enable
If we enable buffering before TLS handshake, there is a chance that
handshake messages will be buffered, such that `tls.Handshake()` can not
read correct payloads from connection.

We add a manual API to control when we will enable buffering. Now
buffering is enabled after the completion of TLS handshake.

Signed-off-by: xhe <xw897002528@gmail.com>